Moved counter and timer components into components folder. Fixed a bug with timer preventing hours greater than 23 from rendering properly.
