+++ src/styles/Paraflow.module.css
@@ -1,6 +1,5 @@
 @import url("https://fonts.googleapis.com/css2?family=Inter&family=Quicksand:wght@600&display=swap");
 
-
 /* Styles for counter component */
 .counterContainer {
   margin: 10px 0;
@@ -26,6 +25,7 @@
   width: 100%;
   height: 100%;
   border: none;
+  cursor: pointer;
 }
 
 .counterNumberDisplay {
@@ -80,6 +80,7 @@
   background-color: #5347D7;
   height: 50px;
   width: 50px;
+  cursor: pointer;
 }
 
 .resetButton {
@@ -87,6 +88,7 @@
   background: none;
   color: #20159E;
   font-size: 50px;
+  cursor: pointer;
 }
 
 .timer {
@@ -134,6 +136,7 @@
   font-size: 32px;
   font-weight: 600;
   padding: 10px;
+  cursor: pointer;
 }
 
 .startButton {
@@ -143,4 +146,5 @@
   width: 100px;
   background-color: #5347D7;
   color: white;
+  cursor: pointer;
 }
